Starting Node.js support
Unfortunately, Jsonary has a lot of global state (listeners, cache,
etc.).

The basic plan (atm) is to create a JS bundle from Jsonary and all the
renderers.  Node-side, this will produce a function to be evaluated on
each request, giving a fresh Jsonary instance.  It can also produce a
bundled JS file for client-side, so you only have to include one JS
file from the HTML to get Jsonary + the same renderer set.